feat: 統一提出 useDisplay hook
